refactor: 변수 의존성 관련 리뷰 반영
