+++ app/src/main/java/com/into/websoso/ui/notification/component/NotificationsContainer.kt
@@ -9,8 +9,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.into.websoso.core.common.util.clickableWithoutRipple
 import com.into.websoso.core.designsystem.theme.WebsosoTheme
+import com.into.websoso.data.repository.NotificationRepository.Companion.DEFAULT_INTRINSIC_ID
 import com.into.websoso.domain.model.NotificationType
-import com.into.websoso.domain.usecase.GetNotificationListUseCase.Companion.DEFAULT_INTRINSIC_ID
 import com.into.websoso.ui.notification.model.NotificationUiModel
 
 private const val LOAD_THRESHOLD = 5
